+++ src/main/java/com/hexagonal/courseregistration/user/UserBeanConfig.java
@@ -2,9 +2,9 @@ package com.hexagonal.courseregistration.user;
 
 import com.hexagonal.courseregistration.user.adapter.persistence.JpaUserAdapter;
 import com.hexagonal.courseregistration.user.adapter.persistence.JpaUserRepository;
-import com.hexagonal.courseregistration.user.application.CheckExistUserPort;
 import com.hexagonal.courseregistration.user.application.RegisterUserUseCase;
-import com.hexagonal.courseregistration.user.application.SaveUserPort;
+import com.hexagonal.courseregistration.user.port.CheckExistUserPort;
+import com.hexagonal.courseregistration.user.port.SaveUserPort;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
