--- src/main/java/com/hexagonal/courseregistration/user/application/CheckExistUserPort.java
@@ -1,5 +1,0 @@
-package com.hexagonal.courseregistration.user.application;
-
-public interface CheckExistUserPort {
-  boolean check(String idNumber, Authority authority);
-}
