--- src/main/java/com/hexagonal/courseregistration/user/application/SaveUserPort.java
@@ -1,5 +1,0 @@
-package com.hexagonal.courseregistration.user.application;
-
-public interface SaveUserPort {
-  void save(User user);
-}
