+++ src/main/java/com/hexagonal/courseregistration/user/port/CheckExistUserPort.java
@@ -1,0 +1,7 @@
+package com.hexagonal.courseregistration.user.port;
+
+import com.hexagonal.courseregistration.user.application.Authority;
+
+public interface CheckExistUserPort {
+  boolean check(String idNumber, Authority authority);
+}
