+++ src/main/java/com/hexagonal/courseregistration/user/port/SaveUserPort.java
@@ -1,0 +1,7 @@
+package com.hexagonal.courseregistration.user.port;
+
+import com.hexagonal.courseregistration.user.application.User;
+
+public interface SaveUserPort {
+  void save(User user);
+}
